Added main view
This creates a set of files for us to work off of.

* corestore.ts will centralize all access to supposedly-server data.  We can jam fake data in here for now to unblock UI development.
* MainChrome.tsx presents the main site chrome.  Haven't integrated it into LoginScreen because we might want that one screen to look different.
* Added react-tabs to support basic tabs.  The idea here is that one person could play many roles, in which case we'd want the app to haev an easy way to flip between that person's tasks, sorted by role (one role per tab).

Everything looks nasty right now, but it's because I've spent zero time on styling.  Perhaps a more important question is what we want to do about CSS in React;  in RN, CSS can be easily embedded into each component's file using StyleSheet.create().  But I don't see an equivalent in React.  For now, I've put individual CSS files using the same name as the class that consumes them;  but the problem is that these styles are technically app-wide (meaning you can easily get into name collisions across components, which smells bad).  Here are a few proposals:

* Do we want a /styles/ directory?  For now, keeping the CSS in the same place as the component feels better... but I could see that feeling messy over time.
* Do we potentially want just one big CSS, so that name collisions are obvious and avoidable (if we alphabetize)?
* Or perhaps do we want a naming convention, like "every component's styles should all fall under a parent style for that component, like .myComponent .myStyle"

LMK thoughts.