Adds CSV upload capability
The UI looks nasty, and has zero investment in it so far because I anticipate we'll change it.  Just introduces an icon which, when clicked, uses an Input object to upload a CSV to Firebase Storage.

This is a temporary solution, as the backend of the project is coming together.  Uploading to Firebase Storage allows us to next use Storage Triggers (as Cloud Functions) to process each upload into a DB (which is also temporary, until our backend is up and running).