Pull auditor tasks from server
We can now pull auditor tasks directly from the server, since CSV upload is working.  As we develop Details View a bit more, we'll need to pull more and more data into the actual ClaimTask;  but for now, this replaces our test data with data that came from a CSV.

I haven't updated the Payor view with this data because we'd likely want to enable a real edit flow instead of doing that (i.e. we'll likely want to implement the "click Approve" functionality in order to populate the Payor view with real data).