Show other active users
In a nod to GDocs, this is a proposal for how we deal with multiple active users working in the same role.  TaskList now writes a little piece of data into Firestore whenever a user clicks on a Task.  Every TaskList subscribes to that collection of active task views, and highlights tasks which others are looking at via an on-hover tooltip.  An orange rectangle highlights any task that someone else is looking at simultaneously (within the last 5 mins).

Note that we don't bother deleting stale ActiveTasks from the Firestore collection.  This is assumed to be ok with our very small user base.  At some later point, we can write some sort of cloud function that cron-jobs itself to remove stale ActiveTasks from the collection.