Introduce remote config and make payments configurable
Remote Config isn't available on React firebase, and so we implement our own super-basic version of it by storing a "metadata/remoteConfig" object in firestore.  We cache reads of that object, and then provide a `getConfig()` function to load that configuration.

I've also changed things so that the Payor tab by default doesn't allow real issuing of payments.  The button label changes, and the payment-issuing functionality is skipped if real payments aren't enabled.