Prevent deduplication of CSV uploads
Introduces a remote config `allowDuplicateUploads` which defaults to false.  When CSVs are uploaded, the first row is checked against previous uploads, and the upload is rejected if the same ID was found previously.

Because the cloud function runs async from when the user uploads the file, we track this type of rejected duplication in AdminLogEvent objects, which will in a later PR be shown in the Admin panel.