Deduplicate uploaded records
This implements full deduplication of CSV uploads on a row-by-row basis by using meta:instanceID from ODK as the unique identifier of records.

* We simplify and flatten the "uploaded records" structure so that one row for each record ever uploaded is stored in the `uploaded_records` collection.
* We scan that collection for newly uploaded meta:instanceIDs, and avoid reinserting IDs we've seen in the past.
* We regenerate a single admin log for each case where a CSV repeats a previously-seen ID.

Note that the "did I previously upload this" search is linear, leading to an O(n^2) implementation of de-duping when it comes to firestore searches.  This is because firestore does not support the concept of querying for records where any of a set of IDs are looked for.  An alternative, faster implementation would be to store all record IDs in one array, since Firestore has an `array-contains` query capability;  however, firestore arrays are limited to a certain modest size, so we wouldn't want to hit that limit.