Update URL history with permalinks
Keep the browser location updated with permalinks whenever you select tabs or tasks so that copy/paste is enabled between users/uses.  Removed class member that tracked whether to respect `initialSelectedTaskID` in favor of replicating prop into state (which is apparently React's #2 recommendation of how to handle that issue).  This also removed the need to re-key MainView.  Added `baseUrl` to each tab's config.  Changed tab name to `Completed` in order to match the case of `Rejected`.